Move token length step definition into token file and delete handlers/auth
