+++ service/src/features/step_definitions/token.js
@@ -1,4 +1,5 @@
-const {When} = require('cucumber');
+const {Then, When} = require('cucumber');
+const {expect} = require('chai');
 const jwt = require('jsonwebtoken');
 
 const isTokenActive = token => {
@@ -19,4 +20,11 @@ When('I get a token', async function() {
 
   const result = await this.utils.request({body, method: 'POST', url});
   this.token = result.body.token;
+});
+
+Then('returned token lasts for thirty minutes', function() {
+  const tokenExpirationInSeconds = jwt.decode(this.activeRequest.body.token).exp;
+  const dateInSeconds = Date.now() / 1000;
+  const thirtyMinutesInSeconds = 60 * 30;
+  expect(tokenExpirationInSeconds).to.equal(dateInSeconds + thirtyMinutesInSeconds);
 });
